Refactored the model builder extensions
* Regrouped to similar extension classes.
* Added missing file headers
* Added missing class comments
* Corrected comments of the methods
