+++ src/Persistence/EntityFramework/Extensions/ModelBuilder/GameMapDefinitionExtensions.cs
@@ -1,14 +1,21 @@
-﻿namespace MUnique.OpenMU.Persistence.EntityFramework.Extensions.ModelBuilder;
+﻿// <copyright file="GameMapDefinitionExtensions.cs" company="MUnique">
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace MUnique.OpenMU.Persistence.EntityFramework.Extensions.ModelBuilder;
 
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using MUnique.OpenMU.Persistence.EntityFramework.Model;
 
+/// <summary>
+/// Extensions for the <see cref="EntityTypeBuilder{GameMapDefinition}"/>.
+/// </summary>
 internal static class GameMapDefinitionExtensions
 {
     /// <summary>
-    /// Extension for GameMapDefinition entity.
-    /// centralize logic for apply configuration on database with code first.
+    /// Applies the settings for the <see cref="GameMapDefinition"/> entity.
     /// </summary>
+    /// <param name="builder">The builder.</param>
     public static void Apply(this EntityTypeBuilder<GameMapDefinition> builder)
     {
         builder.HasMany(map => map.RawEnterGates);
