+++ src/Persistence/EntityFramework/Extensions/ModelBuilder/LetterExtensions.cs
@@ -1,23 +1,30 @@
-﻿namespace MUnique.OpenMU.Persistence.EntityFramework.Extensions.ModelBuilder;
+﻿// <copyright file="LetterExtensions.cs" company="MUnique">
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace MUnique.OpenMU.Persistence.EntityFramework.Extensions.ModelBuilder;
 
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using MUnique.OpenMU.Persistence.EntityFramework.Model;
 
+/// <summary>
+/// Extensions for the <see cref="EntityTypeBuilder{LetterBody}"/> and <see cref="EntityTypeBuilder{LetterHeader}"/>.
+/// </summary>
 internal static class LetterExtensions
 {
     /// <summary>
-    /// Extension for LetterBody entity.
-    /// centralize logic for apply configuration on database with code first.
+    /// Applies the settings for the <see cref="LetterBody"/> entity.
     /// </summary>
+    /// <param name="builder">The builder.</param>
     public static void Apply(this EntityTypeBuilder<LetterBody> builder)
     {
         builder.HasOne(body => body.RawHeader);
     }
 
     /// <summary>
-    /// Extension for LetterHeader entity.
-    /// centralize logic for apply configuration on database with code first.
+    /// Applies the settings for the <see cref="LetterHeader"/> entity.
     /// </summary>
+    /// <param name="builder">The builder.</param>
     public static void Apply(this EntityTypeBuilder<LetterHeader> builder)
     {
         builder.Ignore(header => header.ReceiverName);
